Reorganize code structure
