+++ cmd/channel-manager/types.go
@@ -1,0 +1,10 @@
+package main
+
+import "github.com/luiscarbonell/lightning-node-tools/pkg/lnd"
+
+// Type aliases for convenience - use shared types from pkg/lnd
+type Channel = lnd.Channel
+type ChannelFeeReport = lnd.ChannelFeeReport
+type FeeReportResponse = lnd.FeeReportResponse
+type ForwardingHistory = lnd.ForwardingHistory
+type ForwardingEvent = lnd.ForwardingEvent
